Remove some unused CoreLib stuff
- plRefCnt duplicates hsRefCnt (Wtf?)
- hsNamedPipeStream nasty, unused
- VDB stuff in hsStream left over from Plasma 1 (realMyst/DIRT)
- hsColor32 moved to be less promiscuous
