+++ Sources/Plasma/CoreLib/hsSTLStream.cpp
@@ -153,241 +153,3 @@ const void *hsVectorStream::GetData()
     else
         return nil;
 }
-
-/////////////////////////////////////////////////////////////////////////////////////////
-
-#ifdef HS_BUILD_FOR_WIN32
-
-hsNamedPipeStream::hsNamedPipeStream(uint8_t flags, uint32_t timeout) :
-    fFlags(flags),
-    fPipe(INVALID_HANDLE_VALUE),
-    fReadMode(false),
-    fTimeout(timeout)
-{
-    memset(&fOverlap, 0, sizeof(OVERLAPPED));
-    fOverlap.hEvent = CreateEvent(NULL, TRUE, TRUE, NULL);
-}
-
-hsNamedPipeStream::~hsNamedPipeStream()
-{
-    CloseHandle(fOverlap.hEvent);
-    fOverlap.hEvent = INVALID_HANDLE_VALUE;
-}
-
-bool hsNamedPipeStream::WaitForClientConnect()
-{
-    // Look for a client connect (this should return zero since it's overlapped)
-    BOOL ret = ConnectNamedPipe(fPipe, &fOverlap);
-    if (ret)
-        return true;
-    else
-    {
-        switch (GetLastError())
-        {
-        // Waiting for client to connect
-        case ERROR_IO_PENDING:
-            if (WaitForSingleObject(fOverlap.hEvent, fTimeout) == WAIT_OBJECT_0)
-                return true;
-            break;
-
-        // Client is already connected
-        case ERROR_PIPE_CONNECTED:
-//          if (SetEvent(fOverlap.hEvent))
-                return true;
-            break;
-        }
-    }
-
-    return false;
-}
-
-bool hsNamedPipeStream::Open(const char *name, const char *mode)
-{
-    wchar_t* wName = hsStringToWString(name);
-    wchar_t* wMode = hsStringToWString(mode);
-    bool ret = Open(wName, wMode);
-    delete [] wName;
-    delete [] wMode;
-    return ret;
-}
-
-bool hsNamedPipeStream::Open(const wchar_t *name, const wchar_t *mode)
-{
-    if (wcschr(mode, L'w'))
-    {
-        fReadMode = false;
-
-        // Try to create the pipe
-        fPipe = CreateNamedPipeW(name,
-                                PIPE_ACCESS_OUTBOUND | FILE_FLAG_OVERLAPPED,
-                                PIPE_TYPE_BYTE,
-                                1,
-                                1024,
-                                1024,
-                                fTimeout,
-                                NULL);
-
-        if (fPipe != INVALID_HANDLE_VALUE)
-            return true;
-    }
-    else if (wcschr(mode, L'r'))
-    {
-        fReadMode = true;
-
-        fPipe = CreateFileW(name,
-                            GENERIC_READ, 
-                            0,              // no sharing 
-                            NULL,           // no security attributes
-                            OPEN_EXISTING,  // opens existing pipe 
-                            FILE_FLAG_OVERLAPPED,              // default attributes 
-                            NULL);          // no template file
-
-        if (fPipe != INVALID_HANDLE_VALUE)
-            return true;
-    }
-
-    return false;
-}
-
-bool hsNamedPipeStream::Close()
-{
-    if (fPipe == INVALID_HANDLE_VALUE)
-        return false;
-
-    if (fReadMode)
-    {
-        CloseHandle(fPipe);         // Close our end of the pipe
-        fPipe = INVALID_HANDLE_VALUE;
-    }
-    else
-    {
-        FlushFileBuffers(fPipe);    // Make sure the client is done reading
-        DisconnectNamedPipe(fPipe); // Disconnect the pipe from the client
-        CloseHandle(fPipe);         // Close our end of the pipe
-        fPipe = INVALID_HANDLE_VALUE;
-    }
-
-    return true;
-}
-
-bool hsNamedPipeStream::ICheckOverlappedResult(BOOL result, uint32_t &numTransferred)
-{
-    // Read/Write succeeded, return now
-    if (result)
-        return true;
-    // Read failed because the operation is taking a while.  Wait for it
-    else if (GetLastError() == ERROR_IO_PENDING)
-    {
-        if (WaitForSingleObject(fOverlap.hEvent, fTimeout) == WAIT_OBJECT_0)
-        {
-            BOOL oResult = GetOverlappedResult(fPipe, &fOverlap, (LPDWORD)&numTransferred, FALSE);
-            if (oResult)
-                return true;
-            hsAssert(oResult, "GetOverlappedResult failed");
-        }
-        else
-            hsAssert(0, "Wait failed");
-    }
-    else
-        hsAssert(0, "Read/Write failed");
-
-    return false;
-}
-
-bool hsNamedPipeStream::IRead(uint32_t byteCount, void *buffer, uint32_t &numRead)
-{
-    numRead = 0;
-
-    if (fPipe != INVALID_HANDLE_VALUE && fReadMode)
-    {
-        BOOL result = ReadFile(fPipe, buffer, byteCount, (LPDWORD)&numRead, &fOverlap);
-        if (ICheckOverlappedResult(result, numRead))
-            return true;
-    }
-
-    // If we got here, the pipe is probably broken.  Throw if it is enabled.
-    if (fFlags & kThrowOnError)
-        throw this;
-
-    return false;
-}
-
-bool hsNamedPipeStream::IWrite(uint32_t byteCount, const void *buffer, uint32_t &numWritten)
-{
-    numWritten = 0;
-
-    if (fPipe != INVALID_HANDLE_VALUE && !fReadMode)
-    {
-        BOOL result = WriteFile(fPipe, buffer, byteCount, (LPDWORD)&numWritten, &fOverlap);
-        if (ICheckOverlappedResult(result, numWritten))
-            return true;
-    }
-
-    // If we got here, the pipe is probably broken.  Throw if it is enabled.
-    if (fFlags & kThrowOnError)
-        throw this;
-
-    return false;
-}
-
-uint32_t hsNamedPipeStream::Read(uint32_t byteCount, void *buffer)
-{
-    uint32_t totalRead = 0;
-
-    // Read until we get all our data or an error
-    uint32_t numRead = 0;
-    while (IRead(byteCount-totalRead, (void*)((uint32_t)buffer+totalRead), numRead))
-    {
-        totalRead += numRead;
-
-        if (totalRead >= byteCount)
-            return totalRead;
-    }
-
-    return totalRead;
-}
-
-uint32_t hsNamedPipeStream::Write(uint32_t byteCount, const void *buffer)
-{
-    uint32_t totalWritten = 0;
-
-    // Write until we get all our data or an error
-    uint32_t numWritten = 0;
-    while (IWrite(byteCount-totalWritten, (const void*)((uint32_t)buffer+totalWritten), numWritten))
-    {
-        totalWritten += numWritten;
-
-        if (totalWritten >= byteCount)
-            return totalWritten;
-    }
-
-    return totalWritten;
-}
-
-#ifdef __SGI_STL_PORT
-using std::min;
-#endif
-
-void hsNamedPipeStream::Skip(uint32_t deltaByteCount)
-{
-    char buf[256];
-
-    // Read until we get all our data or an error
-    uint32_t totalRead = 0;
-    uint32_t numRead = 0;
-    while (IRead(min((uint32_t)256L, deltaByteCount-totalRead), buf, numRead))
-    {
-        totalRead += numRead;
-
-        if (totalRead >= deltaByteCount)
-            return;
-    }
-
-}
-
-void hsNamedPipeStream::Rewind()
-{
-    hsAssert(0, "Rewind not allowed on a pipe");
-}
-
-#endif // HS_BUILD_FOR_WIN32
