+++ Sources/Plasma/CoreLib/hsSTLStream.h
@@ -82,48 +82,3 @@ public:
     // In case you want to try and be efficient with your memory allocations
     void Reserve(uint32_t bytes) { fVector.reserve(bytes); }
 };
-
-#ifdef HS_BUILD_FOR_WIN32
-
-
-
-class hsNamedPipeStream : public hsStream
-{
-protected:
-    HANDLE      fPipe;
-    OVERLAPPED  fOverlap;
-    bool         fReadMode;  // True for read, false for write
-    uint8_t       fFlags;
-    uint32_t      fTimeout;
-
-    bool ICheckOverlappedResult(BOOL result, uint32_t &numTransferred);
-    bool IRead(uint32_t byteCount, void *buffer, uint32_t &numRead);
-    bool IWrite(uint32_t byteCount, const void *buffer, uint32_t &numWritten);
-
-public:
-    enum { kThrowOnError = 1 }; // Throws if a read or write operation fails
-
-    hsNamedPipeStream(uint8_t flags=0, uint32_t timeout=INFINITE);
-    virtual ~hsNamedPipeStream();
-
-    // The server (writer) and client (reader) need to open the same file.
-    // The format is "\\.\pipe\pipeName".  The '.' can be replaced with a
-    // computer name to do it over the network.  'pipeName' is whatever you
-    // want.
-    virtual bool      Open(const char *name, const char *mode);
-    virtual bool      Open(const wchar_t *name, const wchar_t *mode);
-    virtual bool      Close();
-    
-    virtual uint32_t  Read(uint32_t byteCount, void *buffer);
-    virtual uint32_t  Write(uint32_t byteCount, const void *buffer);
-    virtual void      Skip(uint32_t deltaByteCount);
-    virtual void      Rewind();
-
-    // - For the server (writer) only -
-    // After calling open, signal your client to start reading and call this function.
-    // If a client connects, this will return true and you can start writing.  If it
-    // returns false, close the pipe, it ain't happening.
-    bool WaitForClientConnect();
-};
-
-#endif // HS_BUILD_FOR_WIN32
